Create Convert to Hexadecimal
You take n and take the base (16) modulo, add to the result string by converting the modulo to digit, and divide n by the base (16). Taking the modulo gives us the last digit in hexadecimal, and by dividing it 'removes' the last digit so we can calculate the second last digit in hexadecimal by repeating the process until n reaches 0.
The time complexity is O(log n), and the space complexity is also O(log n).