Create Remove Adjacent Duplicate Characters
We could compute the final value as described, to iterate and find any adjacent duplicate characters, delete them and repeat. However in the worst case this would give us O(n^2) time, since we need to iterate through all characters in the string each time.

Instead we can use a stack to help. We can iterate through the input string, and if the top of the stack has the same character, it means there is a duplicate and we should pop the top of the stack, otherwise append the character to the stack. This helps us with cases like abba.

The time complexity is O(n), and the space complexity is also O(n).